pyproject.toml
[project]
name = "keyword-spam-collab"
version = "0.1.0"
description = "Helpers and tooling for the Colab-based keyword spam workflow"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "google-cloud-storage>=2.17",
    "pandas>=2.1",
    "numpy>=2.0",
    "scikit-learn>=1.4",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
colab = [
    "torch>=2.3",
    "transformers>=4.44",
    "accelerate>=0.34",
    "datasets>=2.20",
    "trl>=0.11",
    "unsloth>=2024.10.7",
    "bitsandbytes>=0.43 ; platform_system == 'Linux'",
    "peft>=0.12",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["depop"]
py-modules = ["utils"]
